% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartown.R
\name{bartown}
\alias{bartown}
\title{Bar Town}
\usage{
bartown(df, vr = "demographics_member_sex", place = "place",
  maintxt = paste(gsub("_", " ", vr), " per ", place, sep = ""),
  coll = c("red", "darkgreen"), lt = levels(df[, vr]), fn = paste(vr,
  place, "bar.pdf", sep = ""), graphdir = grafdir, ab = TRUE,
  ablevel = 0.5, out = TRUE, ypos = 1.15)
}
\arguments{
\item{df}{Data frame containing personal data}

\item{vr}{Character vector containing the variable (column) under consideration}

\item{place}{Character vector of the place under consideration}

\item{maintxt}{Character vector containing the title of the plot}

\item{coll}{Character vector containing lable colours}

\item{lt}{Legend text used in barplot}

\item{fn}{Character vector containing file name of output plot}

\item{graphdir}{Directory for graphs to be stored in}

\item{ab}{Logical to add one or more straight lines through the current plot}

\item{ablevel}{The y-value(s) for horizontal line(s)}

\item{out}{Logical that creates a copy of the graphics in pdf format to be stored in the desired location}

\item{ypos}{Numeric specifying the y axis position of the legend}
}
\description{
Generic bar diagram for 2 x c table with proportion-test and binomial confidence interval
}

