% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.opsie.tabel.R
\name{capwords}
\alias{capwords}
\title{Capitalize Words}
\usage{
capwords(s, strict = FALSE)
}
\arguments{
\item{s}{Character vector to be converted to upper case}

\item{strict}{Logical that reverts rest of the characters (other than first character) to lower case}
}
\description{
Help function used to capitalize the first letter of a word, also converting following characters
to lower case if specified
}

