% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi.chaid.R
\name{chi.chaid}
\alias{chi.chaid}
\title{Chi2 CHAID}
\usage{
chi.chaid(x, naam = "HH_chi2_uitkomste", datadirr = datadir,
  grafdirr = grafdir, tabdirr = tabdir, p.cutofff = 0.01, plott = TRUE,
  nett.naam = "HH_chi2_uitkomste.net")
}
\arguments{
\item{x}{Data frame containing survey information}

\item{naam}{Character vector containing the name attributed to the output data}

\item{datadirr}{Character vector referring to the location of the output directory for generated data}

\item{grafdirr}{Directory in which graphs will be stored as Character vector}

\item{tabdirr}{Directory in which tables will be stored as Character vector}

\item{p.cutofff}{The cutoff probability value as a Numeric}

\item{plott}{Logical whether to plot data or not}

\item{nett.naam}{Character vector containing the name of the gml network diagram}
}
\description{
Pairwise Chi2 tests together with CHAID
}
\details{
Relies on chi.2.exp and spdf function
}

