% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onttrek.hh.indikatore.R
\name{CI}
\alias{CI}
\title{CI}
\usage{
CI(x, ci = 0.95, na.rm = TRUE)
}
\arguments{
\item{x}{Data frame containing numeric data}

\item{ci}{The confidence interval as a numeric}

\item{na.rm}{Logical that removes NA values when TRUE}
}
\description{
Creates statistical informative output for numeric data based on a confidence interval
}

