% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decodeMobenzi.R
\name{decodeMobenzi}
\alias{decodeMobenzi}
\title{Decode Mobenzi}
\usage{
decodeMobenzi(dfSurvey = NULL, dfCodeBook = NULL,
  fldnmVariable = "Variable", fldnmValue = "Value", fldnmLabel = "Label")
}
\description{
Decode Mobenzi
}

