% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eentweedrie.R
\name{eentweedrie}
\alias{eentweedrie}
\title{One, Two & Three}
\usage{
eentweedrie(x = qol, drops = c("sol_waste_notcollected_inside_recycled",
  "sol_waste_notcollected_outside_other",
  "sol_waste_notcollected_outside_other", "sol_income_sources_explainother",
  "respondent_info_contact_phonenumber"), een = TRUE, eendir = tabdir,
  twee = TRUE, tweedir = paste(tabdir, "tweerigting/", sep = ""),
  n.tweedir = paste(tabdir, "n.tweerigting/", sep = ""),
  driedir = paste(tabdir, "drierigting/", sep = ""), groep = "place",
  groeplab = NULL, moederskip = TRUE, p.type.n = "box", graph = FALSE,
  graphdir = "~/Dropbox/Sasol QOL/Working/Analysis/QualityOfLifeSurvey/Report/Grafieke/",
  verbose = FALSE, ...)
}
\arguments{
\item{x}{Data frame containing survey data}

\item{drops}{Character vector of the names of columns to be dropped}

\item{een}{Logical that generates one way LaTeX tables when TRUE}

\item{eendir}{Character vector of the directory in which the one way LaTeX tables should be saved}

\item{twee}{Logical that generates two way LaTeX tables when TRUE}

\item{tweedir}{Character vector of the directory in which the two way LaTeX tables should be saved}

\item{n.tweedir}{Character vector of the directory in which the numeric two way LaTeX tables
should be saved}

\item{driedir}{Character vector of the directory in which the three way LaTeX tables should be saved}

\item{groep}{Character vector that indicates the grouping variable}

\item{groeplab}{The label of the grouping variable as character vector}

\item{moederskip}{Logical that creates tables with additional information}

\item{p.type.n}{The plot type as character vector}

\item{graph}{Logical that creates and saves a qplot when TRUE}

\item{graphdir}{Character vector of the directory in which the graphs should be saved}

\item{verbose}{Logical that displays function messages}

\item{...}{Arguments that are passed to/from other functions}
}
\description{
Function for one, two and three way LaTeX tables and graphs
}

