% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkIndex.R
\name{hkIndeks}
\alias{hkIndeks}
\title{HK Index}
\usage{
hkIndeks(hk, dimensies = 3, cuts = c(0.75, 0.7, 0.65), verbose = FALSE)
}
\arguments{
\item{hk}{Object of type PCA}

\item{dimensies}{Number of dimensions as Numeric}

\item{cuts}{Numeric vector that describe acny cuts to be made (uts should be the length of the number of
dimensions for most reliable output)}

\item{verbose}{Logical to display fuction messages}
}
\description{
Makes an inex out of PCA
}
\details{
use: hkIndeks(swb.pca, cuts = c(75, 70, 65) ,  dim = 3)
}

