% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hkIndex.R
\name{hkIndeksPerd}
\alias{hkIndeksPerd}
\title{HK Index Perd}
\usage{
hkIndeksPerd(hk, cuts = c(75), dim = 1)
}
\arguments{
\item{hk}{Object of type PCA}

\item{cuts}{Numeric vector of the cuts to be made}

\item{dim}{Numeric vector containing the number of dimensions}
}
\description{
Help function for HK Index
}

