% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.formula.c2m.R
\name{make.formula.c2m}
\alias{make.formula.c2m}
\title{Make Formula c2m}
\usage{
make.formula.c2m(x, add = NULL, verbose = FALSE)
}
\arguments{
\item{x}{An spdf}

\item{add}{Anything to be added to the formula as character vector}

\item{verbose}{Logical that displays function messages when TRUE}
}
\description{
Formula maker for the Chi2 machine
}

