% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.opsie.tabel.R
\name{multi.opsie.punt}
\alias{multi.opsie.punt}
\title{Multi Option Point}
\usage{
multi.opsie.punt(df = des, x = "energy_lpg_utility", yescat = "Yes",
  nocat = "No", lm = 0.02, abs.perc = TRUE)
}
\arguments{
\item{df}{Data frame that contains survey data}

\item{x}{Variable under consideration as a character vector}

\item{yescat}{Character vector indicating "yes" category}

\item{nocat}{Character vector indicating "no" category}

\item{lm}{Lower limit that takes output into consideration when percentage of responders are larger
than proposed limit}

\item{abs.perc}{Logical that includes percentage of total results information when TRUE}
}
\description{
Creates a tex output that describes the percentages of people who chose variables (will only display if
percentage is larger than a specified one)
}

