% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.opsie.tabel.R
\name{multi.opsie.tabel}
\alias{multi.opsie.tabel}
\title{Multi Option Table}
\usage{
multi.opsie.tabel(df = des, x = "energy_lpg_utility", yescat = "Yes",
  nocat = "No", tabeleer = TRUE, abs.perc = FALSE, tex = TRUE,
  tabdir = get("tabdir", envir = .GlobalEnv), graphdir = get("grafdir",
  envir = .GlobalEnv), fn = paste(x, "h.bar.", sep = ""), verbose = FALSE,
  fplace = "!ht")
}
\arguments{
\item{df}{Data frame. Contains quality of life survey information}

\item{x}{Character vector. The variable that should be isolated}

\item{yescat}{Character vector indicating "yes" category}

\item{nocat}{Character vector indicating "no" category}

\item{tabeleer}{Logical whether to create LaTeX table or not}

\item{abs.perc}{Logical that includes percentage of total results information when TRUE}

\item{tex}{Logical that prints LaTeX table when TRUE}

\item{tabdir}{Directory in which tables will be stored as Character vector}

\item{graphdir}{Directory in which graphs will be stored as Character vector}

\item{fn}{File name to be attributed to saved tables as Character vector}

\item{verbose}{Logical That displays function messages when TRUE}

\item{fplace}{Table placement as character vector}
}
\description{
Function creates graphs displaying multiple choice results for specified variables
}

