% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.prev.table.R
\name{pop.prev.table}
\alias{pop.prev.table}
\title{Population Prevalence Table}
\usage{
pop.prev.table(data = persoon, age = "demographics_member_age",
  sex = "demographics_member_sex",
  disease = "body_symptoms_twelvemonth_list_wheezing.and.tight.chest.at.night.or.after.exercise",
  binwidth = 10, verbose = FALSE, drop.na = TRUE, outname = TRUE,
  show.margins = FALSE)
}
\arguments{
\item{data}{Data frame that contains personal data}

\item{age}{Character vector referring to member age column}

\item{sex}{Character vector referring to member sex column}

\item{disease}{Character vector of the disease under consideration}

\item{binwidth}{Numeric that is used for calculating the number of intervals into which x is to be cut}

\item{verbose}{Logical that displays function messages when TRUE}

\item{drop.na}{Logical that drops NA values when TRUE}

\item{outname}{Logical that will assign the disease name to the global environment when TRUE}

\item{show.margins}{Logical vector of variable names
(can include "grand\_col" and "grand\_row") to compute margins for,
or TRUE to compute all margins . Any variables that can not be margined over will be silently dropped}
}
\description{
Creates PPT that can also serve as input to a population pyramid
}
\details{
Can then use pyramid.plot(res$Male_Yes,res$Female_Yes,labels=res$Age_group, main=disease, gap=1.5, show.values=TRUE), OR
pyramid.plot(res$Male_1,res$Female_1,labels=res$Age_group, main=NULL, gap=50, show.values=TRUE)
}
\examples{
res = pop.prev.table(disease=NULL)
dorp.p <- by(data=persoon,INDICES = persoon$place, FUN = function(x) pop.prev.table(data=x, disease=NULL))
}

