% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pos.negify.R
\name{pos.negify}
\alias{pos.negify}
\title{Pos Negify}
\usage{
pos.negify(df, negs = c("Strongly disagree", "Disagree"), pos = c("Agree",
  "Strongly agree"))
}
\arguments{
\item{df}{Data frame containing multiple choice data}

\item{negs}{Character vector referring to negative connotations}

\item{pos}{Character vector referring to positive connotations}
}
\description{
Links "positive" and "negative" to answers indicating "agree" and "disagree"
}

