% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.table.R
\name{report.table}
\alias{report.table}
\title{Report Table}
\usage{
report.table(x, caption, label, directory)
}
\arguments{
\item{x}{The object you wish to export}

\item{caption}{Character vector of length 1 or 2 containing the table's caption or title}

\item{label}{Character vector of length 1 containing the the LaTeX label}

\item{directory}{Character vector containing the directory to which the table must be saved}
}
\description{
Function to export a LaTeX table with caption and label
}

