% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swb.bar.R
\name{swb.bar}
\alias{swb.bar}
\title{SWB Barplot}
\usage{
swb.bar(data, var, group, plot = TRUE, teks = TRUE, middel = 5,
  graphdir = "~/", verbose = TRUE)
}
\arguments{
\item{data}{Data frame containing survey data}

\item{var}{Character vector that contains swb variables}

\item{group}{Character vector referring to a grouping variable}

\item{plot}{Logical whether to create a plot (not present in this function)}

\item{teks}{Logical whether to create a LaTeX output (not present in this functon)}

\item{middel}{Numeric. Values smaller than this value will have proportion tests done on them for
each town.}

\item{graphdir}{Character vector containing the graph output directory}

\item{verbose}{Logical to display function messages}
}
\description{
Creates a SWB barplot Pos - Neg per town
}

