% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swb.bar.R
\name{swb.line}
\alias{swb.line}
\title{SWB Line}
\usage{
swb.line(m, group = "place", facet = FALSE, uit = TRUE,
  graphdir = "~/tmp/", uitnaam = "SWBline", outform = ".pdf",
  yfix = FALSE)
}
\arguments{
\item{m}{A melted data frame}

\item{group}{Grouping variable as character vector}

\item{facet}{Logical}

\item{uit}{Logical that saves the graph in the graphdir when TRUE}

\item{graphdir}{Directory to which graphs must be saved}

\item{uitnaam}{Name to be attributed to the output graph as character vector}

\item{outform}{The type of output desired as character vector}

\item{yfix}{Logical (Not present in this function)}
}
\description{
Creates an SWB line plot per town. Takes a melted data frame as input
}

