% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swb.toets.groep.R
\name{toets.groep}
\alias{toets.groep}
\title{Test Group}
\usage{
toets.groep(data = qol, x = "swb_satisfaction_general", group = "place",
  group2 = NULL, tabdir = "~/tmp/", graphdir = "~/tmp/", middel = 5)
}
\arguments{
\item{data}{Data frame containing survey data}

\item{x}{The swb variable under consideration as character vector}

\item{group}{Character vector containing the grouping variable}

\item{group2}{Character vector containing an optional second grouping variable}

\item{tabdir}{Directory for tables to be saved in}

\item{graphdir}{Directory for graphs to be saved in}

\item{middel}{Numeric. Values smaller than this value will have proportion tests done on them for
each town. Bonferroni's method is used to calculate p-values}
}
\description{
Test differences in swb tables per group
}

